backend/requirements.txt
fastapi[all]
pymongo
passlib[bcrypt]
python-jose[cryptography]
python-dotenv